feat(resource)!: B+ ResourceFactory contract, retire ResourceDescriptor (ADR-0095 D2)
Make the resource arm of the plugin contract symmetric to the action arm.

- Introduce object-safe `ResourceFactory` (introspection: key/metadata/validate;
  construction: `register(&Manager, RegisterRequest) -> BoxFut<SlotIdentity>`).
- Move `ResourceActivator`/`KindActivator`/`RegisterRequest` down from
  nebula-engine into nebula-resource (`factory.rs`); engine re-exports them.
- Retire describe-only `ResourceDescriptor` (no shim/alias).
- `#[derive(Resource)]` emits a per-R factory instantiating `KindActivator`,
  topology author-spelled via a new `#[topology(Resident|Pooled)]` attribute.
- `Plugin::resources()` returns `Vec<Arc<dyn ResourceFactory>>`.
- Disambiguate the colliding `register_resource` -> `register_resource_action`.
- Three frozen laws hold: schema-single-source, removal-funnel, key-coherence.

Latent by design: the `register` arm has zero production callers until
bind-population (M12.4); slots land `SlotIdentity::Unbound`. F6 (async_trait)
and bind-population are out of scope.

Verified: cargo check --workspace --all-targets, clippy -D warnings,
nextest (764 passed), rustdoc -D warnings — all green.
