+++ crates/engine/Cargo.toml
@@ -58,7 +58,7 @@ uuid = { workspace = true, features = ["v4", "serde"] }
 humantime-serde = "1.1"
 lru = { workspace = true }
 parking_lot = { workspace = true }
-scopeguard = "1"
+scopeguard = { workspace = true }
 regex = { workspace = true }
 url = { workspace = true }
 futures = { workspace = true }
